+++ src/server/devices/baseDeviceInterfaces/iHeater.ts
@@ -5,6 +5,7 @@ export interface iHeater extends iRoomDevice {
   settings: HeaterSettings;
   desiredTemperature: number;
   readonly humidity: number;
+  // The level between 0 and 1.0
   readonly iLevel: number;
   readonly iTemperature: number;
   roomTemperature: number;