+++ src/server/devices/zigbee/zigbeeTuyaValve.ts
@@ -50,7 +50,7 @@ export class ZigbeeTuyaValve extends ZigbeeHeater {
     switch (idSplit[3]) {
       case 'valve_position':
         this.log(LogLevel.Trace, `Tuya Valve valve_position Update for ${this.info.customName} to "${state.val}"`);
-        this._level = state.val as number;
+        this._level = (state.val as number) / 100;
         break;
       case 'local_temp':
         this.log(LogLevel.Trace, `Tuya Valve Local_Temp Update for ${this.info.customName} to "${state.val}"`);